+++ extensions/index.ts
@@ -1,6 +1,9 @@
 export { AligningGuidelines } from './aligning_guidelines';
 export type * from './aligning_guidelines/typedefs';
 
+export { Scrollbars } from './scrollbars';
+export type * from './scrollbars/typedefs';
+
 export {
   originUpdaterWrapper,
   installOriginWrapperUpdater,